+++ core/QuickForm2.php
@@ -10,15 +10,13 @@
 namespace Piwik;
 
 use HTML_QuickForm2;
-use HTML_QuickForm2_InvalidArgumentException;
 use HTML_QuickForm2_Node;
-use HTML_QuickForm2_NotFoundException;
 use HTML_QuickForm2_Renderer;
 
 /**
  * Manages forms displayed in Twig
  *
- * For an example, @see Piwik\Plugins\Login\FormLogin
+ * For an example, @see \Piwik\Plugins\Login\FormLogin
  *
  * @see                 HTML_QuickForm2, libs/HTML/QuickForm2.php
  * @link https://pear.php.net/package/HTML_QuickForm2/
@@ -64,8 +62,6 @@ abstract class QuickForm2 extends HTML_QuickForm2
      * @param    mixed $attributes Element attributes
      * @param    array $data Element-specific data
      * @return   HTML_QuickForm2_Node     Added element
-     * @throws   HTML_QuickForm2_InvalidArgumentException
-     * @throws   HTML_QuickForm2_NotFoundException
      */
     public function addElement(
         $elementOrType,